Move serialize module and flake8
